Split reducer/actions into three modules (projects, view, log)
