+++ src/projects/index.ts
@@ -1,0 +1,4 @@
+import reducer from './reducer'
+import * as actions from './actions'
+
+export { reducer, actions }
