+++ src/view/index.ts
@@ -1,0 +1,4 @@
+import * as actions from './actions'
+import reducer from './reducer'
+
+export { actions, reducer }
